messages: Extract tests for messageReducers from chatReducers tests.
The recent "messages: Normalize `narrows` state`" extracted some
features that previously resided in chatReducers.js into a new file
messageReducers.js.  So, extract to an appropriate new file all the
tests for chatReducers that actually test features that are now in
messageReducers.

A handful of tests (all at the top of the new file) are copied without
deleting from the old file, because they'll be useful for both files.
For now we skip them too, to keep the bulk move simple.

In the next commit, we'll fix up all these tests for the new state
structure and to refer to `messageReducers` rather than
`chatReducers`, and re-enable them.
